refactor: clean modules
